[INTERNAL][FIX] ObjectPageLayout: Fixed scrolling issues on IE11
Problem: While we scroll nested controls of the ObjectPageLayout with
the scroll-wheel there is incorrect display of their scroll container.

Root Cause: When ObjectPageLayout scroll container has the
"sapUxAPObjectPageWrapperTransform" CSS class the container itself has
the CSS property "transform" set to "translate3d(0, 0, 0)". This
sometimes causes issues on other scroll containers of nested controls.

Solution: Since the support for the "transform" CSS property is limited
and can be considered volatile as in the current case, we set it
to "none" on Internet Explorer browser.

Change-Id: I3d3f75e45dbf1c9ead4027beb4b38e553deabc0e
BCP: 0020751294 0000432437 2019
BCP: 1970472626