[INTERNAL] Table: Change scroll algorithm to index + offset information
The scroll position is now an index + offset value. Compared to a scroll
position represented as a floating-point number, the calculations become
more reliable as fewer rounding errors are likely to occur. The offset
can be a pixel value or a percentage. When scrolling with the scrollbar,
an offset is set in percent first. An offset in pixels is necessary for
exact calculations and UI updates and will therefore be set as early as
possible as part of the update process.

Updates of the scroll position that are triggered by various actions,
e.g. setting the "firstVisibleRow" property or scrolling the scrollbar,
are implemented as cancellable processes. An update process is aborted
if an update process of at least the same rank is triggered. For
example, an update triggered by scrolling the viewport has the lowest
rank and can only cancel processes that were triggered by scrolling the
viewport, but no other processes.

All tests that where commented out in the QUnit tests for vertical
scrolling could be commented in. More tests are added to cover even more
scenarios.

Change-Id: I3ae497a1c37bf7bf4b237657de12c662e3d5803a