[INTERNAL][FIX] DynamicPage: Header and title shadow fixed on Fiori 3
Problem: By Visual Design there are supposed to be shadows below the
header/title when there is no additional navigation
(IconTabBar etc.) and vice-versa.

Solution: We introduce the CSS class "sapFDynamicPageNavigation"
which we attach to the ObjectPageLayout control when there is
additional navigation. This class serves as a helper for the addition
and the removal of the above mentioned shadows when needed.

Change-Id: I74fe98f71664b60fcf9aa7cb88ff9a8057084266
BCP: 1970512557