[INTERNAL] Table: Suppress reset when calling unbindAggregation for rows
The reset should be suppressed when calling "unbindAggregation" for the
rows aggregation. Reset calls "destroyAggregation" -> "destroyRows". But
the table does not allow to call this method and logs an error. A reset
with destroyAggregation is also not required for the rows aggregation,
because the table manages the rows by itself.

Change-Id: Ic4b957f6865f663bd436ce4e223a499938c484d4