[FEATURE] sap.uxap.ObjectPageHeader: New background color property introduced
Previously we were using sap.m.Icon as objectImage when there is an
icon (valid icon URI for the objectImageURI) or a
placeholder (showPlaceholder is true).

Now we are using sap.f.Avatar in those two cases.

This allows us to introduce the newly added "objectImageBackgroundColor"
property which decides the background color for the objectImage in the
cases of icon or a placeholder.

The possible values for the new property are described in the
"sap.f.AvatarColor" enum.

Additionally a bug is fixed where if you try to set the shape or
background color of a placeholder after its rendering, the placeholder
disappears.

Change-Id: I62ec99dc3ff5dacef3d5f959c4920358141de028
BCP: 1970466184
JIRA: BGSOFUIPIRIN-4241