[INTERNAL][FIX] Table: Correctly leave edit mode on row header cell
In edit mode, the ItemNavigation is disconnected from the keyboard
navigation of the table. If the table switches from edit to navigation
mode when the focus is on a row header cell, there is no focus change
and the item navigation needs to be notified about the currently focused
cell manually.

Change-Id: I2a55925b69fcdc0ae0c0c6fa08ca104b1ad6c0d1