[FIX] sap/base/Log: Limit amount of stored log entries
The Log module stores all log messages for the current log-level
in an array. The stored log entries were never deleted or limited.
To avoid memory or performance issues, the stored entries are now
limited to 3000 entries. When the limit is reached, 30 percent of the
currently stored log entries are discarded. An API to change the limit
is provided (e.g. for the SupportAssistant, OPA tests).
The provided listeners are informed via onDiscardLogEntries when the
log entries are being discarded.

Fixes: https://github.com/SAP/openui5/issues/2689
Change-Id: I963065fc2afe6a00f62b0ba1fda2bb0a130ab546