[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Header shadows fixed on Fiori 3
Problem: By Visual Design there are supposed to be shadows below the
header/title when there is no additional navigation
(AnchorBar, IconTabBar etc.) and vice-versa.

Solution: We introduce the CSS class "sapUxAPObjectPageNoNavigation"
which we attach to the ObjectPageLayout control when there is no
additional navigation. This class serves as a helper for the addition
and the removal of the above mentioned shadows when needed.

Change-Id: I766bc149c43df4924d69d078d43ab5dfa7a38f4a
BCP: 1970512557