[INTERNAL][FIX] ui5loader: delay the switch to microtasks
Many tests seem to be badly written as they don't explicitly wait for
the UI5 Core to have booted, but they use functionality that is only
available after the Core has been initialized.

With this change, the loader uses tasks by default, but the Core
switches it to use microtasks, once the Core has initialized.

Change-Id: I02f7e19d692043ee2796a124c1a6c4762342733d
BCP: 1970540094