[INTERNAL] XMLTreeModifier: triggering destroy after removal fixed
When an aggregated child control is removed during XMLProcessing (e.g.
in a flex change), the node is removed from it's parent. So a subsequent
destroy (which may not be needed in this case) fails.
However to make using modifiers easy for consumers, the behaviour should
be similar to how things happen at runtime (JS).
One can then just call:
    oModifier.removeAggregation(oControl, "columns", oColumn);
    .. followed by
    oModifier.destroy(oColumn);
without having to know about "xmlTree" or "jsControlTree"

Change-Id: I9f06620ae406fa6bef150e9c44e21576c6eb1468