[INTERNAL][FIX] Table: Scrollbar scrolling if re-rendered in hidden DOM
If the table is in auto mode and re-rendered while invisible (e.g. if
the parent DOM element is invisible), the internal rows updated event is
not fired. The event should then be fired when the table is made
visible, otherwise necessary updates of the table will not be performed.

Change-Id: I278b72457e73cf549db1c3951e5b14fdf6347d4f
BCP: 1970532705