[INTERNAL][FIX] sap.m.Toolbar: Fixed cutting of items
Problem:
First item is cut from the left side when there
is an item which is not visible.

Root cause:
If the item is hidden, it doesn't have .sapMBarChild
class and this is not the last child of its parent.

Fix:
Added additional selector that matches the case when
there is a hidden element and checks which item of the toolbar
is the last visible one.

Change-Id: If3462193cd819684767b83c07e1405dab63dd8f3
BCP: 1980378210