[INTERNAL][FIX] sap.f.DynamicPage: Fixed incorrect RTA configuration
Problem:
When we collapse and then expand the header with a click,
instead of a scroll, the content in the header is no longer editable
during RTA.

Root Cause:
When the above mentioned steps are applied, the header is no longer
part of the content (scroll) wrapper as stated in the designtime file.
It starts being part of the title wrapper which is never a scroll container.

Solution:
When we are in a state, in which the header is part of the title (sticky)
container, and not the content (scroll) container, we have the private
flag "_bHeaderInTitleArea" being set to true. We are using this boolean
as a helper in the function which returns the aggregations which are part
of the scroll container. If we have it set to true, we return only the
content aggregation as a part of the scroll container.

In theory it should be always true if the DynamicPage is pinned, which
makes part of the check in the scroll container function (_bPinned)
redundant. I am leaving it for eventual backward incompatibility.

Change-Id: I2ac8bad42daeb0a85112c21a5b76350a2b37f0f3
BCP: 1970427657