[INTERNAL][FIX] sap.uxap.ObjectPageSubSection: Updated documentation
Problem: When we have more than one block in the blocks aggregations
and one of them is with the "sapUxAPObjectPageSubSectionAlignContent"
CSS class, there is an incorrect alignment.

Root Cause: Negative margins are applied to the blocks with the
"sapUxAPObjectPageSubSectionAlignContent" CSS class, which interferes
with the correct alignment of the rest of the blocks (if any).

Solution: We updated the documentation regarding the
"sapUxAPObjectPageSubSectionAlignContent" CSS class. Now we state that
the App developer should avoid using this class if there is more than
one control in the same ObjectPageSubSection.

Change-Id: Ia80940bdd184130db5445661a78e3397e48c19e4
BCP: 1880591703