[FIX] sap.uxap.ObjectPageHeader: Fixed element checking
Appanrently jQuery does not return correct result when counting
visible elements and overflow button gets hidden.
This is due to incorrect reference used to retrieve the buttons -
it is different than the reference containing the overflow button.
This itself is due to cloning of elements and using them for measurements.

Now the same header DOM reference is used.

Change-Id: I9d42dc89de889378389d7359cc23697b0d4bf390
BCP: 1970569747