[INTERNAL] sap.m.InputBase: Handle concurrency between user input and data model
The change introduces the _setPreferUserInteraction restricted method. It is used
to set the preferred user interaction - if set to true, overwriting
the user input with model updates will be prevented.
Also, when the user has focused on an input field, but did not type anything, the
new value coming from the model will be applied to the field, but it will get
selected/highlighted to make it possible for the user to easily overwrite it.

Change-Id: Ie76d70180a027b3715dcf4bcd71c5943c93f5b5a
JIRA: BGSOFUIRILA-2600