[INTERNAL] m.Table: navigated column fix
When the "Delete" or "SingleSelect" mode is enabled on the table, the
navigated column header was rendered before the column headers for the
modes mentioned above. Due to this mismatch of column headers and cells,
the delete button and the single select radio button were not interactive
any more. The column header are now aligned with the cell content and the
issue is fixed. Additionally, some CSS clean-up.
BCP: 1970579923

Change-Id: Ifef808a7d0cfd43a0fe8d2a727cc28187d2fc0f9