[FIX] sap.ui.unified.Calendar: Now the year range button is not cut
There are a lot of cases for which the year button is used:
- for a single year in one calendar
- for a single year in multiple calendars
- for a range of years in one calendar
- for a range of years in multiple calendars

This change fixes the case where there is a range of years in one
calendar without braking the others.
Previously the standalone Calendar scenario was, and it's still working,
because it added a css class during _initializeHeader and the
corresponding selector was doing its job. This is way too
early for the case in this change where the Calendar is used inside a
PlanningCalendar.

Change-Id: I17de8fec8efcc659c9ae74183b06fefaec6cb3f6
BCP: 1970573833
BCP: 1980454204
Fixes: https://github.com/SAP/openui5/issues/2725