[FIX] sap.m.MultiInput: Improved performance on addToken
We don't need to execute and sync all the relations between
Token and Selected List items at bootstrap. Usually, it's required on
user interaction. Further more, the List items are not actually
needed until the User opens the popover.
Not delaying the execution causes performance penalties which
might be observed at tables and when there are too many items selected
in the MultiInput.

BCP: 1970531093

Change-Id: Iae45c16ca3ac295c0b6f9004d23297c866a01b6b