[INTERNAL] sap.m.SimpleFixFlex: Flex container now rendered with id
- After the implementation of the SimpleFixFlex control in the SuggestionsPopover
we have encountered the issue that the currently available implementations for
scrolling to the selected item no longer works, as the control that is
responsible for the scroll is no longer the Popover.
This fix will provide us the means to correctly select the part of the DOM
that is responsible for the scroll.

Change-Id: I2dbc0795226ebbacb8cbb0776ec3ae762b265a44