[FIX] sap.m.MultiInput: Duplicate tokens on mobile fixed.
When there are no keys in the tokens, on mobile devices
they seem to be duplicated. The duplicate tokens themselves are not an
issue- tokens without keys should be an invalid scenario; there's
no other "unique" field which could be used and combination of fields
is rather ambiguous and not sustainable.

However, duplication in the case of mobile devices seems to be coming from
duplicate execution of code which could be omitted.

BCP: 002075129400002353582020

Change-Id: Ib1e2f86f09606128bce2e37c76d6be5b353d3d92
CR-Id: 002075125800001456692020