[FIX] sap.m.Slider: Tickmarks evenly rendered
Although calculations for the distance between tickmarks is very precise,
some of the browsers are not able to render properly such a precision-
especially numbers with infinite exponent (3.333333333%, for example).

We'd need to either round those numbers or just hide the tickmarks
followed by labels. In ideal case, where calculations are
properly interpreted, the tickmarks and labels collide.
Rounding the numbers would be tough to implement and with unclear end result.

Change-Id: I93c2eeb93f7c285b157fc3b43f33b07223c16f38
BCP: 1970472623