[FIX] Demo Kit: Prevent reduntand page reload
Problem:
In the APIRef section, in the "detail" page of a selected entity,
when the user is scrolling between the sections,
a perceived reload occurs whenever the URL is changed

This happens because the URL change triggers popstate event
and the popstate event listener explicitly tells the router
to load the view that corresponds to the new URL (without
checking first if that page is already displayed)

Solution:
Fixed for now by extending the existing previous fix that
prevents loading the same view more than once.
A more central solution will be researched later.

Change-Id: I6964f4d45538fd10af4fb2aa06222e213b2ff096
BCP: 1970597228