[FIX] sap.f.FlexibleColumnLayout: ObjectPage's header flickering is fixed
Problem: The Header of the sap.uxap.ObjectPageLayout is flickering when
navigating from OneColumn Layout of the sap.f.FlexibleColumnLayout to
TwoColumn layout. The reason: a wrong width value is measured after the
resuming of the ResizeHandler because there is a CSS class toggling
immediately after the resuming, which, in some cases, changes the width.

Solution: First toggling the CSS class and after that resuming the
ResizeHandler, so that the width of each Column is correct.

Change-Id: I9eb55a593a072cbdf198dde984e78a5bd6deea0d
BCP: 1980354228