[FIX] sap.m.MultiInput: Trigger resize handler on token update
This is an edge case where the token's text gets updated from the outside.
In that scenario, we'd need to invalidate the MultiInput, so
all calculations are properly done.

However, tokenUpdate in the MultiInput is impossible due to the
forwardAggregation functionality. The solution is to listen for Tokenizer
resize in the DOM and then to trigger the resize handler.

Change-Id: I9720fcf8be4322f310e794d7d3ab721b59fba568
BCP: 1970502405