[FIX] sap.ui.unified.Calendar: Belize plus colors adjusted
Belize plus works in this way - it sets contrast to the header of the
container and renders all the items inside with contrast as well.
The rest of it remains light. That's why if a Calendar is inside of the
light part, it has to apply the belize light theme. Unless it has a
sapContrastPlus class explicitly set. Then and only then it has to take
the colors from the belize plus file.
Before this change it was taking the Legend's colors for working/non
working day everytime when on belize_plus without caring if there is a
contrast class.

Change-Id: Ic785f04cb3a68738f11fc10c0006840e67a8f379
CR-Id: 002075125900005553852019
BCP: 002075129400006570372019
(cherry picked from commit 42e5ae61e5c2cd3a88764e0592bd3b546c89ee84)