[INTERNAL] sap.m.Wizard: Now implements sap.f.IDynamicPageStickyContent interface
- Wizard's progress navigator is now sticked to the DynamicPage's sticky header.
- In this scenario, two things need to be taken into account:
   1. The DynamicPage's fitContent property needs to be set to 'true' so that the
scroll is handled by the Wizard itself. This way the Wizard's functionality to scroll
to its active steps is preserved.
   2. 'sapUiNoContentPadding' class should be added to the DynamicPage
as the Wizard should take 100% of the width.
- The two points above were added into the documentation.
- For better look and feel responsive paddings should be added to the
content and the header areas of the Wizard.

- There is one limitation though: As the DynamicPage does not have scroll anymore,
the floating footer does not go over the Wizard, but is left underneath it and behaves
like normal footer.

Change-Id: I0f596607f964fffa1bef5b65e53a24d6e150b04a
JIRA: BGSOFUIRILA-2583