[INTERNAL][FIX] sap.ui.rta Don't break stack initialization
Stack initialization is not build with control variants or
app descriptor changes in mind. But when those appear in the workspace
the initialization logic was breaking.
This change will make the stack ignoring change files it doesn't know.
Having support for control variants or app descriptor changes is up to
another change.

Change-Id: I46b93ac93f60cbb61b767b53ee886e2635e33c2b