[INTERNAL] m.Table: fix focus outline with sticky column headers
In the recent change where the gap between the background color and border
outline was fixed with I1486d853df8ab269aa58ff6630eef8317d308434, caused
another issue with sticky. The focus outline on the table header row was
not visible on all 4 sides. Hence, adding the background-clip on the <th>
element when the header row is focused. This ensure the bottom focus outline
is also visible when the browser is zoomed.
BCP: 1970632972

Change-Id: Ieaa78464e15161265e93ae5053366982dd05dba4