[FIX] sap.uxap.AnchorBar: Unnecessary focus is now removed
Problem: When we focus the arrow down of the sap.uxap.AnchorBar
there are two focuses. One on the icon (arrow down square) and one on
the whole clickable area. There should be focus only on the square
with the down arrow.

Root Cause: There is a pseudo element which comes from the sap.m.Select
styles, which causes the unwanted focus.

Solution: We force the pseudo element's (focus) outline style to be
"none".

Additionally a bug is fixed where when you hover on the clickable area
of the down arrow, an unwanted border is shown.

Change-Id: I222e6c3df18170ef30e0e2c0a8911ba7c87f6ace
BCP: 1970611339