[INTERNAL] jsdoc-template: fixes to improve consistency of api.json
# Return Type of getMetadata

The generated documentation for UI5 classes (created with *.extend)
uses a fixed return type 'sap.ui.base.Metadata' for method
'getMetadata'. But classes that enhance metadata with their own concepts
(like ManagedObject, Element, Component, Controller etc.) usually
document a more specific return type for 'getMetadata'. For subclasses
of those specializations, this results in inconsistent overrides of
'getMetadata'.

With this change, the best matching public return type is chosen. This
is done in a post-processing step, because, at the point in time when
the auto-doc is initially generated, the necessary knowledge about the
inheritance tree is not yet complete.

# Suppress Module Export for type definitions and interfaces

The information about the module export name does not make sense for
interfaces and typedefs as they don'T have a runtime representation.
Instead of displaying an error, the export information is now completely
suppressed.

# Basename

In some cases, JSDoc does not provide a valid basename in property
symbol.name, but a partially qualified name. The UI5 template now
ensures an unqualified basename.

Change-Id: I294a57006bfed3ed3812fb34cd26687f3addb895