[INTERNAL] Core: make dnd.DragSession an interface
DragSession was documented as a namespace with a global name. But
there's no global object with that name. And over time, many different
instances of DragSession are created, which also doesn't match the
semantics of a namespace. An interface is more appropriate.

Documentation has been enhanced with a note clarifying the nature of the
interface.

Change-Id: I90f332a59569908d2341383ce07edd84a4d30e87