[INTERNAL][FIX] sap.f.FlexibleColumnLayout: Performance optimization
Problem: The width of the control is always measured from the DOM, which
triggers a lot of unneeded layout recalculations.

Solution: As the width can be changed only at two places - in
onAfterRendering method (when the control is rendered for the very first
time) and in _onResize method (called when there is a resize event), we
measure the width only there and cache the value. The resize handler of
the FlexibleColumnLayout is never getting suspended (only the resize
handlers of the columns are suspended during the animation), so the
caching of the control's width cannot lead to any problem. When caching
the value, we prevent most of the layout recalculations, which
significantly improves the performance.

Change-Id: Id2cd1bc58b33a67a16b43a236366cc7cb2c14940
JIRA: BGSOFUIPIRIN-4262