[INTERNAL] sap.ui.integration: refactor propertyChange event
Previously all property editors were attached directly to BaseEditor
and provided their value directly to the context model. This works
well for root editors (those which are defined in the BaseEditor
configuration) and doesn't work when we want to compose several
property editors one into another. In this scenario, cascading take
place and an outer property editor should be reposible for output
value. Thus, nested editors from now on do not communicate their
values directly to the BaseEditor, but rather inform the outer
property editor about the change and the outer property editor
decides what to do next with it.

Change-Id: Idcc207449eb3f0bc36e9345c4155ab537dc98d34