[INTERNAL][FIX] sap.m.CheckBox: Fixed visual misalignments
Problem:
The Label of CheckBox is not displayed correctly if the CheckBox
is the only child of the Toolbar. It overlaps with the CheckBox
and it's displayed under it.

Root cause:
The left padding of the CheckBox is overwriten with a newly
added stronger selector coming from the Toolbar.

Fix:
The previous rule for the correct padding is now applied
for the stronger selector coming from the Toolbar.

Change-Id: I71c1d931b89e537a34e7bf6b93ff33c423bee16f