[FEATURE] sap.m.InitialPagePattern: New sample added
The initial page floorplan allows the user to navigate to a single
object to view or edit it. The interaction point on the screen is a
single input field that relies on assisted input to direct the user to
the object in as few steps as possible (using features such as value
help and live search).

The floorplan is achieved with a single sap.uxap.ObjectPageLayout with
sap.uxap.ObjectPageDynamicHeaderTitle.

sap.m.Input is used as an Input within the Title.
sap.m.MessagePage is used as a section, which informs the user
if input is needed or if the searched value is not found.
sap.m.SelectDialog is used as a value help dialog for the input.

Change-Id: Ie140d83c51ab6e4fbba7f132fb85d393948ae118
JIRA: BGSOFUIPIRIN-4322