[FIX] sap.f.DynamicPage: Scrolling with large header on Desktop is fixed
Problem: When the expanded header is too big the page cannot be scrolled
down on Desktop.

Solution: There is logic, which overrides "preserveHeaderStateOnScroll"
property, when the header is too big on Tablet and Mobile devices, which
makes scrolling possible. This logic is now adjusted and works the same
way for Desktop too. The Documentation is also changed.

Change-Id: I10381c1e6532d790fda6694d716310ebd07b75c4
BCP: 1970554221