[FIX] sap.m.ScrollBar: Racing condition issue with setContentSize fixed
The problem: When we change the property contentSize of sap.m.ScrollBar
the change causes re-rendering of the control, which in combintation
with setting the scrollPosition property, which is done synchronously is
causing wrong calculation of the scroll position. When there is still no
height set to the content of the control, the scrolling is being
positioned on the top of the container, which is losing the previous
scroll position. In the case we found it, in sap.f.DynamicPage in IE 11
we are unsnapping the header not intentionally, which was breaking some
OPA tests.

example: Snippix #87661
(run in IE11, version >= 1.70.0, in new tab with screen size apx. 800px)

Solution: We need to also add custom setter for the contentSize
property, as it is for the scrollPosition, and suppress invalidation of
the control itself in order to force the order of performing DOM
changes.

BCP: 1980348279

Change-Id: I41f99398809b14dbcb08bf85aeb8babe3bc8cf59