[FIX] sap.m.PlanningCalendar: Doesn't load density aware image by default
Previously PlaningCalendarRow uses the default behavior of the
StandardListItem and tried to load the density aware image.
However, this causes multiple 404 requests for the applications
that do not have these images.

Consider adding a property in PlanningCalendarRow for this in the future
if requests on this comes.

BCP: 002075129400007126932019

Change-Id: I2d258ea5ba5e50058d6a4340605733af1301e87b
(cherry picked from commit e6c75ee5df05014ffc014bc3ddbfa67f70fa6a5e)
CR-Id: 002075125900005974852019