[INTERNAL] sap.uxap.ObjectPageLayout: Buttons IDs placed on MenuItems
Problem: Hierarchical select was not showing any subsections because
created buttons are destroyed and new MenuItem controls are created
in their place.

Solution: Menu items are given the IDs of the previous buttons so they can be
target when selectedButton is about to change. Old buttons are created
with autogenerated IDs.

Change-Id: I8e917097beb6d0adb0410163067eeec051ec2200
JIRA: BGSOFUIPIRIN-4276