[INTERNAL] Table: Enrich row with state information
2nd step to prepare for a plugin that enables usage of
functionality provided by the V4 ODataListBinding (CPOUIFTEAMB-1095).
1st step: Ib23f5da9787e6a6532834cfacdbac4818a458428

The row gets state information (e.g. type, level, expandable, ...) that
is updated when a binding context is set. This information is exposed
table internally via private getters. Most code that indirectly
determined the state of a row, e.g. by reading the DOM, is replaced with
this new row API. In the remaining code, there is no easy access to a
row instance.
A hook is available to adjust the row state. This hook is now used by
the TreeTable, AnalyticalTable and the experimental grouping to have a
clearer code separation.

Change-Id: I0b945cb64a54f74caf977fe096b5fa65bc54eb77