[INTERNAL][FIX] sap.m.DraftIndicator: Partly cut label is now visible
Problem: Label is partly cut on its end.

Root Cause: A recently introduced regression where "overflow: hidden" is
added to the label text wrapper. This causes the italic text to be cut
at the end.

Solution: We force "overflow: visible" on the label text wrapper, only
if it's in the context of the DraftIndicator.

Change-Id: I10e5ffe9cee3c5236e4ac624002d91f65ee839a0
BCP: 1980465086