[INTERNAL] Table: Add menu hooks
3rd step to prepare for a plugin that enables usage of
functionality provided by the V4 ODataListBinding (CPOUIFTEAMB-1095).
1st step: Ib23f5da9787e6a6532834cfacdbac4818a458428
2nd step: I0b945cb64a54f74caf977fe096b5fa65bc54eb77

Add one hook before a menu is opened to add menu items to it. Another
hook to inform the column that menu items will be added on the next
opening of the column's menu (necessary for acc, styles, ... -
see Column#_menuHasItems).
The AnalyticalTable is making use of the "OpenMenu" hook to provide its
group header menu items.

Change-Id: Ifd4f69072b2142635f8f7c36e8c234add867ba29