[FIX] sap.f.FlexibleColumnLayout: Measuring width when FCL is visible
Problem: When the FCL is not visible, because it has a parent with
display=none, a wrong width is measured. If setLayout function is
called in the meanwhile, before the ResizeHandler is triggered, due to
a change in parent's visibility, incorrect width value is used.

Solution: Any time when the _getControlWidth function is called and
the width value is still 0, the width is getting measured from the
DOM if the control is visible. This way the width is measured correctly
on time, when there is a change in parent's visibility.

Change-Id: I2702c81d419e2328ae267cf43c7f6a226c356e7c
BCP: 1980497298