[INTERNAL][FIX] Table: Rows aggregation rebind behavior
If the rows are already bound, it is part of the rebind process of the
ManagedObject to first unbind the aggregation. Do not react to this
unbind action (unbindAggregation) to avoid unnecessary DOM updates and
UI flickering.
It is the responsibility of the consumer to interact with rows according
to their state, and block the UI for the end user if necessary -
optionally assisted by the auto busy handling of the table.

Change-Id: I5a1802148daff16fc9abc0838f8922a9d2d81bc1
BCP: 1980497568