[FIX] sap.m.PlanningCalendar: Now the button's year text is correct
Previously, when open a Calendar inside a PlanningCalendar, navigate to
the year picker, close it and reopen it, there was an issue. The button,
which shows the year had a wrong text - the previous range of years
instead of only one year.
That was because nobody ensured that the picker mode will be the correct
one.
Now it's handled as it is in the DatePicker - on a popup's 'closed' event.

Ensured by a visual test which simulates the scenario above.

Change-Id: I38702ca148a5477f12b02f0a39298108eac07fc3
BCP: 1980459813