[INTERNAL][FIX] sap.f.DynamicPage: Fixed z-index of pseudo element
Problem: Left pseudo element of the pin and toggle
header indicator buttons is not displayed properly.

Background: A previous solution led to regression
where a sticky header is above ::before
element of the toggle indicator and hides it.

Solution: We remove the z-index logic which led to the regression.
Also we are not displaying the left pseudo element of the pin
button on Edge browser.

Change-Id: Icd6b4dd2e2dea1799a688ffb38bff44a8131cdca
JIRA: BGSOFUIPIRIN-4276