[FIX] sap.f.shellBar.ControlSpacer: Width is set immediately on DomRef
Problem: Sometimes, when ControSpacer's width is changed, a race
condition happens. The doLayout function of OverflowToolbar is executed,
before the ControlSpacer is rerendered and wrong width value is cached.
However, at this moment, the OverflowToolbar has its final width, and
when the doLayout is called again:
- it uses the already cached (wrong value)
- setControlsOverflowAndShrinking is not called, because the width of
the OverflowToolbar is the same as it was on the previous call

Solution: A custom setter of ControlSpacer is introduced for setWidth.
It sets the width immediately on the DomRef. This way, the
OverflowToolbar caches the correct width value and uses it for its
calculations.

Change-Id: I23676f8dd056a650ddbbd86d9071a0fd0fb92d61
BCP: 1970461036