[FEATURE] Demo Kit: Frame removal in developer guide pages
Background: Demo Kit displays developer guide pages in an iframe.
This leads to a number of issues where we cannot benefit from
Google search indexing and also bad theming capabilities.

Since the iframe will no longer be rendered the new html content is
rendered in a sap.ui.core.HTML control. Therefore some changes are
needed in the documentation context:

1. External links are proxied trough a sap disclaimer.

2. Collapsible type sections - expand/collapse function is handled.

2. A subtopic is no longer handled by a second "#" symbol.
This was previously handled in the scope of the iframe,
now there is new route matching the topic/subtopic pattern
and content is scrolled to the subtopic programmatically.

3. All imagemaps in the topic are taken and an ResponsiveImageMap is
created, managing the size of an image and size of clickable areas
and hover styles.

Additionally a synchronous call to the html source is changed to async.

Change-Id: Iaba99c55ccc6bdd42fdc919c1a6d7af4c8677235
JIRA: BGSOFUIPIRIN-4299