[INTERNAL] sap.ui.dt - TaskManager pending task functionality added
With this change the designtime sync event is now chained and waiting
for the overlay applyStyles functionality. The sync event is now fired
after the scrollbars syncing is done and after all applyStyles and Sizes
of relevant overlays are finished. This improves the stability of OPA
and unit tests.

PS-18: mutation on body when overlay-container is added is now ignored.
       That helps in many qunit tests. This body mutation causes two
       DesignTime sync-events after DesignTime initialisation.

BCP: 2080064671

Change-Id: Iab0e7e8e43d9b9d1a5b71e211cd40fbb1ea543e6
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>