[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Failing build is fixed
Problem: For some reason, sometimes when the QUnits are executed, the
DOM tree is not cleaned, and when we search for an element with
sapUxAPAnchorBar class, we find one, but in another rendered ObjectPage.

Solution: Search for AnchorBar only in the ObjectPage, which is rendered
in the currently executed QUnit.

Change-Id: Ica8f26336be473ffdef613836e17ffec46f3d827
BCP: N/A