[INTERNAL] m.Tree: fix longDragOver which fires the toggleOpenState event
When the Tree is configured to be DnD enabled and the tree items are lazily
loaded, the longDragOver only expands the tree node on which the dragged
tree item is over but does not fire the toggleOpenState event. This is now
fixed such that the toggleOpenState event if fired when an node is expanded
via DnD interaction.
Fixes: https://github.com/SAP/openui5/issues/2760
BCP: 1980542394

Change-Id: I1d0f2c3e412eab412e4542bb0febc2f31720e89c