[INTERNAL] sap.m.IconTabBar: Migrated to Semantic Rendering
When there is a tab which has never beed opened (rendered)
and it gets changed (invalidated), the whole IconTabBar gets
invalidated as well. This causes re-rendering of the current active tab,
which can stop events. The issue is reported with sap.m.Text,
which now doesn't have custom setters, after being migrated to
semantic rendering, but this can happen with any control that
is put as a content. Migrating the IconTabBar to
semantic rendering solves this issue.

Change-Id: If710b4ae10cf20be3799c4120f0bead0c40b8e8e
JIRA: BGSOFUIRODOPI-2145
Snippix: 98217
BCP: 1980416461