[INTERNAL] sap.m.SinglePlanningCalendar: stickyMode fixed
When the control is in stickyMode NavBarAndColHeaders (meaning the
navigation toolbar and the column day headers are sticked atop) and is
rendered in Fiori 3 theme, the column headers was positioned wrong.
Now the calulation considers the height of the actons toolbar,
not the height of the navigation one (the actions toolbar will be hidden,
not the navigation one).

Change-Id: Ib1bd11cec48c6d0cd446c9328c4bf1644de89874