[INTERNAL] sap.m.Slider, sap.m.RangeSlider: Tooltip Quartz theme adjustments
-- Slider & RangeSlider tooltip has been modified to follow
the Fiori 3 specifications.
-- Fixed visual bug of the tooltip where the input field in it
covered the top border of the tooltip. This bug was in both
Quartz Light and Dark themes and it is fixed in both of them.

Now the tooltip has a border in order for it to look more
like our input control, but also the Fiori 3 specs for
the tooltip of the Slider & RangeSlider are covered.

Change-Id: I6bb9dbfca8b177a87578553f4355402ee2df5a85
JIRA: BGSOFUIRILA-2664