[FEATURE] Theming: experimental CSS variables and skeleton build/usage
This change enables the usage of CSS variables. The configuration has
been extended for an experimental flag "xx-cssVariables" which supports
the following modes:

 - "true" || "x": pure CSS variables and skeleton CSS per library
 - "additional": loads the CSS variables in addition to the classic
library CSS
 - "false": loads the classic library CSS only (default)

This change requires that all loaded libraries are using the latest
theme build which produces the CSS variables (css-variables.css) and the
skeleton CSS (library-skeleton.css) per used library.

The FOUC handling will also be used for loading the skeleton CSS files
to ensure to avoid flickering on theme switch.

For IE11 the usage of the CSS variables will be suppressed as this
feature is not supported for this browser.

Change-Id: I4dd986cdd70993599f2d8b1459bd9d4f3559b8e0