[FIX] sap.ui.layout.ResponsiveSplitter: Preserve HTML Control content
On each re-rendering of ResponsiveSplitter we destroyed all internal pages.
This caused losing of DOM refs and thus, content in HTML Controls, for example,
was returned to its initial state.

Now we don't destroy all the pages, but add new pages if it is needed,
or destroy only as much as we have to. This way we ensure that the pages
which shouldn't be touched are not touched. Also when destroying, we
preserve the content.

Change-Id: I215babd61883993cf679b8e9f7c935c0230471a9
BCP: 1980546244