[FIX] sap.m.MultiInput: Token expansion prevented on Icon click
Clicking on the ValueState icon, triggers focusin.
However, when there are too many tokens it blocks the UI and the ValueState event does not get fired. We need to postpone Tokens expansion to be the latest thing in the queue.
Further more, the focusin event gets triggered 2 times:
- on icon click, as the Icon is part of the MultiInput
- on the real focus.
So, we could skip the first expansion and leave just one.

BCP: 1970628555

Change-Id: I7dd836db8e03b6e905071cc6ce4195ade3398510