[INTERNAL] FESR: missing bytes received, reset async counter
Only requests that belongs to an Interaction are taken into
account for calculating the received bytes. The passport action
must be the same as the FESR stepName (ignoring the action counter).
This change fixes the comparison so the bytes are calculated
correctly.

When starting a new Interaction the async cvounter/timer should
be resetted.

Change-Id: Id1d7e99b105f361c2f2b60138dea8ec415b41bab