[INTERNAL] Table: Add "_hiddenDependants" aggregation
This private hidden aggregation is similar to the public "dependents"
aggregation. The difference is, besides that it is private, that it
is not cloned. Therefore, it can be used to safely add table internal
instances as children of the table, which also couples their data
binding and lifecycle to the table.

Change-Id: I022efd97f8eaf08ba5dd5b9fddaae93b17fd5375