[FIX] sap.m.Carousel: Changing pages with fast clicks is now possible
If the arrows are clicked fast, now we check if animation has already
started. In such cases call the callbacks that would be usually called
when the animation completes and deregister listeners (this is already
possible with the mobify-carousel, so just call its methods).
Then just call .move, which will start another animation for the next element.

Fixes: https://github.com/SAP/openui5/issues/2783
Change-Id: I193e9686e70a2e924f33c84908d1c77f3496f3b4
BCP: 002075129500006800032019
CR-Id: 002075125900000329172020